renamed container files for consistency
